# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Ebdaa Furniture Factory Management System API

    **Integration (web vs legacy routes):**
    - `factory-hub` paths (`/factory-hub/*`) back the Grand Line React dashboard — JSON-first `fh_*` tables in LibSQL.
    - `/wooden/*` and `/metal/*` remain the stable HTTP contract for ERP-style integrations; they use `wooden_*` / `metal_*` tables.
    - Optional env `FH_SYNC_WOODEN=true` enables a bridge that mirrors hub wood updates into `wooden_work_orders` via a documented field mapping (see `docs/legacy-adapter-factory-hub.md`).
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees with optional filters
      parameters:
        - name: department_id
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: factory_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: redact
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        jobTitle:
          type: string
        standardizedRole:
          type: string
        hireDate:
          type: string
          nullable: true
        departmentId:
          type: string
          nullable: true
        factoryId:
          type: string
        departmentName:
          type: string
          nullable: true
      required:
        - id
        - name
        - jobTitle
        - standardizedRole
        - factoryId
...